Move files to `fixtures` directory
